backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Compress==1.14

# AI and Machine Learning (Python 3.13 compatible)
gpt4all==2.0.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
transformers==4.46.3
huggingface-hub==0.26.2

# PDF Processing
pdfplumber==0.11.4

# Caching & Performance
cachetools==5.5.0
diskcache==5.6.3

# Utilities
python-dotenv==1.0.0
psutil==7.1.0
Werkzeug==3.0.1

# Data Processing (Python 3.13 requires numpy 2.x)
numpy==2.1.3

# Optional: For production
# gunicorn>=21.2.0  # WSGI server for production
